Move tokensizer to shared fixture
